requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
pydantic>=2.7
python-dotenv
python-dateutil
openai>=1.0.0
